requirements.txt
Click==7.0
Flask==1.0.2
Flask-API==1.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Pillow==6.0.0
Werkzeug==0.15.2
